backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
motor>=3.6.0
beanie==1.25.0
pydantic>=2.9.2
pydantic-settings==2.2.1
clerk-backend-api==1.2.0
resend==0.8.0
httpx==0.27.0
cryptography>=43.0.1
python-dotenv==1.0.1
PyJWT==2.9.0
tenacity==8.3.0
pytest==8.1.1
pytest-asyncio==0.23.6

